+++ sw/source/uibase/uiview/pview.cxx
@@ -232,7 +232,7 @@ void SwPagePreviewWin::CalcWish( sal_Int16 nNewRow, sal_Int16 nNewCol )
                               mnSttPage, maPaintedPreviewDocRect );
     SetSelectedPage( mnSttPage );
     SetPagePreview(mnRow, mnCol);
-    mfScale = double(GetMapMode().GetScaleX());
+    mfScale = GetMapMode().GetScaleX();
 
     // If changes have taken place at the columns, the special case "single column"
     // must be considered and corrected if necessary.
@@ -375,7 +375,7 @@ void SwPagePreviewWin::SetWinSize( const Size& rNewSize )
     if ( mbCalcScaleForPreviewLayout )
     {
         mpPgPreviewLayout->Init( mnCol, mnRow, maPxWinSize );
-        mfScale = double(GetMapMode().GetScaleX());
+        mfScale = GetMapMode().GetScaleX();
     }
     mpPgPreviewLayout->Prepare( mnSttPage, Point(0,0), maPxWinSize,
                               mnSttPage, maPaintedPreviewDocRect );
@@ -385,7 +385,7 @@ void SwPagePreviewWin::SetWinSize( const Size& rNewSize )
         mbCalcScaleForPreviewLayout = false;
     }
     SetPagePreview(mnRow, mnCol);
-    mfScale = double(GetMapMode().GetScaleX());
+    mfScale = GetMapMode().GetScaleX();
 }
 
 OUString SwPagePreviewWin::GetStatusStr( sal_uInt16 nPageCnt ) const
@@ -1746,7 +1746,7 @@ void SwPagePreviewWin::AdjustPreviewToNewZoom( const sal_uInt16 _nZoomFactor,
                                   mnSttPage, maPaintedPreviewDocRect );
         SetSelectedPage( mnSttPage );
         SetPagePreview(mnRow, mnCol);
-        mfScale = double(GetMapMode().GetScaleX());
+        mfScale = GetMapMode().GetScaleX();
     }
     else if ( _nZoomFactor != 0 )
     {
